Gather CLI-related stuff into one place
