+++ divik/cli/_utils.py
@@ -12,7 +12,7 @@ import tqdm
 
 import divik.utils as u
 from divik import __version__
-from divik._data_io import load_data
+from divik.cli._data_io import load_data
 
 
 def parse_args():
