+++ divik/cli/divik.json
