+++ divik/cli/inspect.md
