+++ divik/cli/spectral.md
