+++ divik/utils.py
@@ -1,6 +1,7 @@
 from typing import Callable, Tuple, Dict, NamedTuple, List, Optional
 
 import numpy as np
+from skimage.color import label2rgb
 
 Table = np.ndarray  # 2D matrix
 Data = Table
@@ -32,3 +33,15 @@ def normalize_rows(data: Data) -> Data:
     norms = np.sum(np.abs(normalized) ** 2, axis=-1, keepdims=True)**(1./2)
     normalized /= norms
     return normalized
+
+
+def visualize(label, xy, shape=None):
+    x, y = xy.T
+    if shape is None:
+        shape = np.max(y) + 1, np.max(x) + 1
+    y = y.max() - y
+    label = label - label.min() + 1
+    label_map = np.zeros(shape, dtype=int)
+    label_map[y, x] = label
+    image = label2rgb(label_map, bg_label=0)
+    return image
